fix: change parseint to float

+++ src/application.ts
@@ -83,7 +83,7 @@ export class PocketGatewayApplication extends BootMixin(ServiceMixin(RepositoryM
     const influxOrg: string = INFLUX_ORG || ''
     const archivalChains: string[] = (ARCHIVAL_CHAINS || '').replace(' ', '').split(',')
     const alwaysRedirectToAltruists: boolean = ALWAYS_REDIRECT_TO_ALTRUISTS === 'true'
-    const ttlFactor = parseInt(REDIS_LOCAL_TTL_FACTOR) || 1
+    const ttlFactor = parseFloat(REDIS_LOCAL_TTL_FACTOR) || 1
 
     if (aatPlan !== AatPlans.PREMIUM && !AatPlans.values.includes(aatPlan)) {
       throw new HttpErrors.InternalServerError('Unrecognized AAT Plan')
